refactor: charts 디렉토리 구조 변경 및 배당금 차트 컴포넌트 분리
