+++ src/components/Portfolio/Charts/Dividend/DividendBarChartContainer.tsx
