+++ src/components/Portfolio/Charts/Sector/SectorBar.tsx
