+++ src/components/Portfolio/Charts/Sector/SectorBarChartContainer.tsx
