+++ src/components/Portfolio/Charts/Sector/SectorBarItem.tsx
